app/requirements.txt
django
gunicorn
psycopg2-binary
djangorestframework
flake8
twilio